+++ packages/insomnia/src/sync/git/migrations.ts
@@ -26,8 +26,7 @@
 
 import { database } from '~/common/database';
 import { type GitCredentials, type GitRepository, services } from '~/insomnia-data';
-import type ElectronStorage from '~/main/electron-storage';
-import { initElectronStorage } from '~/main/window-utils';
+import { getElectronStorage } from '~/main/electron-storage';
 
 import * as models from '../../models';
 
@@ -35,15 +34,6 @@ const { isGitCredentialsOAuth } = models.gitRepository;
 const { isGitCredentialsV1 } = models.gitCredentials;
 
 const MIGRATION_KEY = 'GIT_CREDENTIALS_MIGRATION';
-
-let electronStorage: ElectronStorage | null = null;
-
-const getElectronStorage = () => {
-  if (!electronStorage) {
-    electronStorage = initElectronStorage();
-  }
-  return electronStorage;
-};
 
 const hasRunMigration = () => {
   const migrationStorage = getElectronStorage();
